Improve the bindings generation process and regenerate them. (#105)
* Install ClangSharpPInvokeGenerator as a local tool and update it.

* Exclude the `dump` functions from being generated.

Now the binding generation does not need any manual steps.
And move more options to generate.rsp.

* Regenerate the bindings.

* ulong -> nuint

* Tweak the bindings generation instructions.

+++ sources/TileDB.CSharp/File.cs
@@ -11,12 +11,12 @@ namespace TileDB.CSharp
     public sealed unsafe class File
     {
         private readonly Context _ctx;
-        
+
         public File(Context ctx)
         {
             _ctx = ctx;
         }
-        
+
         public ArraySchema SchemaCreate(string uri)
         {
             var ms_uri = new MarshaledString(uri);
@@ -28,7 +28,7 @@ namespace TileDB.CSharp
             }
             return new ArraySchema(_ctx,array_schema_p);
         }
-        
+
         public void URIImport(string arrayURI, string fileURI, MIMEType mimeType)
         {
             var ms_arrayURI = new MarshaledString(arrayURI);
@@ -36,7 +36,7 @@ namespace TileDB.CSharp
             var tiledb_mime_type = (tiledb_mime_type_t)mimeType;
             _ctx.handle_error(Methods.tiledb_filestore_uri_import(_ctx.Handle, ms_arrayURI, ms_fileURI, tiledb_mime_type));
         }
-        
+
         public void URIExport(string fileURI, string arrayURI)
         {
             var ms_fileURI = new MarshaledString(fileURI);
@@ -44,50 +44,50 @@ namespace TileDB.CSharp
             _ctx.handle_error(Methods.tiledb_filestore_uri_export(_ctx.Handle, ms_fileURI, ms_arrayURI));
         }
 
-        public void BufferImport<T>(string arrayURI, T value, ulong size, MIMEType mimeType) where T: struct
+        public void BufferImport<T>(string arrayURI, T value, nuint size, MIMEType mimeType) where T: struct
         {
             var ms_arrayURI = new MarshaledString(arrayURI);
             var data = new[] { value };
             var dataGcHandle = GCHandle.Alloc(data, GCHandleType.Pinned);
             var tiledb_mime_type = (tiledb_mime_type_t)mimeType;
             _ctx.handle_error(Methods.tiledb_filestore_buffer_import(
-                _ctx.Handle, 
-                ms_arrayURI, 
+                _ctx.Handle,
+                ms_arrayURI,
                 (void*)dataGcHandle.AddrOfPinnedObject(),
                 size,
                 tiledb_mime_type));
         }
-        
-        public T BufferExport<T>(string arrayURI, ulong offset, ulong size) where T: struct
+
+        public T BufferExport<T>(string arrayURI, nuint offset, nuint size) where T: struct
         {
             var ms_arrayURI = new MarshaledString(arrayURI);
             var data = new T[1];
             var dataGcHandle = GCHandle.Alloc(data, GCHandleType.Pinned);
             _ctx.handle_error(Methods.tiledb_filestore_buffer_export(
-                _ctx.Handle, 
-                ms_arrayURI, 
+                _ctx.Handle,
+                ms_arrayURI,
                 offset,
                 (void*)dataGcHandle.AddrOfPinnedObject(),
                 size));
-            
+
             var result = data[0];
             dataGcHandle.Free();
 
             return result;
         }
-        
+
         public ulong Size(string arrayURI)
         {
             var ms_arrayURI = new MarshaledString(arrayURI);
-            ulong size;
+            nuint size;
             _ctx.handle_error(Methods.tiledb_filestore_size(
-                _ctx.Handle, 
+                _ctx.Handle,
                 ms_arrayURI,
                 &size));
 
             return size;
         }
-        
+
         public string MIMETypeToStr(MIMEType mimeType)
         {
             var tiledb_mime_type = (tiledb_mime_type_t)mimeType;
@@ -96,18 +96,18 @@ namespace TileDB.CSharp
             {
                 Methods.tiledb_mime_type_to_str(tiledb_mime_type, p_result);
             }
-            
+
             return ms_result;
         }
-        
+
         public MIMEType MIMETypeFromStr(string str)
         {
             tiledb_mime_type_t mimeType;
             var ms_str = new MarshaledString(str);
-            
+
             unsafe
             {
-                int status = TileDB.Interop.Methods.tiledb_mime_type_from_str(ms_str, &mimeType); 
+                int status = TileDB.Interop.Methods.tiledb_mime_type_from_str(ms_str, &mimeType);
                 if (status != (int)Status.TILEDB_OK)
                 {
                     throw new System.ArgumentException("MIMETypeFromStr, Invalid string:" + str);
